requirements.txt
telethon
python-dotenv
loguru
sqlalchemy
psycopg2-binary
fastapi
uvicorn
dbt-postgres
ultralytics
dagster
dagster-webserver
